requirements.txt
# Core Web Framework Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
websockets==12.0

# Semantics and Graph Processing
torch==2.6.0
torchvision==0.21.0
torchaudio==2.6.0
transformers==4.51.1
sentence-transformers==2.3.1

faiss-cpu==1.9.0
numpy==1.25.0
scipy==1.15.3
scikit-learn==1.7.1
networkx==3.4.2

# Text Processing
spacy==3.7.5
tiktoken==0.9.0
json-repair==0.46.2

# Document Parsing
magic-pdf==1.3.12
python-docx==1.1.2
pypdf==4.2.0

# API & HTTP Clients
openai==1.102.0
requests==2.32.5

# Configuration & Utilities
PyYAML==6.0.2
nanoid==2.0.0
dotenv==0.9.9 
python-dotenv==1.1.1
chardet==5.2.0